refactor: 리뷰확인 후 코드 수정
