+++ src/main/kotlin/blackjack/domain/GameConditionNotify.kt
@@ -1,5 +1,8 @@
 package blackjack.domain
 
+import blackjack.domain.player.Player
+import blackjack.domain.player.Players
+
 interface GameConditionNotify {
     fun giveDefaultCardsToPlayerDone(players: Players)
     fun isNeedMoreCard(player: Player): Boolean
