+++ src/test/kotlin/blackjack/domain/card/CardTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 import io.kotest.assertions.throwables.shouldNotThrow
 import org.junit.jupiter.params.ParameterizedTest
