+++ src/test/kotlin/blackjack/domain/player/DelearTest.kt
@@ -1,5 +1,8 @@
-package blackjack.domain
+package blackjack.domain.player
 
+import blackjack.domain.card.Card
+import blackjack.domain.card.CardNumber
+import blackjack.domain.card.CardType
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
